requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
ecdsa==0.19.0
fastapi==0.115.5
h11==0.14.0
idna==3.10
pyasn1==0.6.1
pydantic==2.10.1
pydantic_core==2.27.1
python-jose==3.3.0
python-multipart==0.0.17
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
